chapter_6/requirements.txt
absl_py==1.3.0
ale_py==0.7.5
gym==0.25.2
matplotlib==3.7.0
numpy==1.24.3
pandas==1.5.3
